Backend/requirements.txt
asgiref==3.8.1
django~=5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-api-key==2.3.0
djangorestframework_simplejwt==5.5.0
Markdown==3.7
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2

python-dotenv~=1.1.1
stripe~=12.2.0